scripts/05_evaluate_models.py
"""
STEP 05: Model Evaluation
Evaluates all trained models and checks performance targets
"""

import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent.parent))

import yaml
from loguru import logger

# INSERT CODE HERE
# Extract and implement Phase 5 logic from run_full_pipeline.py
#
# Expected implementation:
# 1. Load configuration
# 2. Initialize PerformanceChecker
# 3. Check all models against targets (R² ≥ 0.85 mechanical, R² ≥ 0.80 ballistic)
# 4. Generate evaluation reports
# 5. Save metrics to results/metrics/evaluation_summary.csv
