requirements.txt
# Core Scientific Computing
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# Machine Learning - Tree-based Models
scikit-learn==1.4.0
xgboost==2.0.3
catboost==1.2.3
lightgbm==4.3.0

# Intel Optimizations (CRITICAL for i7-12700K)
scikit-learn-intelex==2024.1.0

# Hyperparameter Optimization
optuna==3.5.0
hyperopt==0.2.7

# Model Interpretation
shap==0.44.1
eli5==0.13.0

# Materials Science
pymatgen==2024.2.8
mp-api==0.41.2
jarvis-tools==2024.2.1
matminer==0.9.0

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
tqdm==4.66.2
joblib==1.3.2

# Logging
loguru==0.7.2

# Statistical Analysis
statsmodels==0.14.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-xdist==3.5.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Jupyter
jupyter==1.0.0
ipykernel==6.29.0
ipywidgets==8.1.1
